+++ tests/NCronJob.Tests/RuntimeJobRegistryTests.cs
@@ -152,6 +152,36 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     }
 
     [Fact]
+    public async Task RemovingByJobTypeDisabledTypeJobsAccountsForAllJobs()
+    {
+        ServiceCollection.AddNCronJob(s => s.AddJob<DummyJob>(p => p.WithCronExpression("1 * * * *")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<DummyJob>(p => p.WithCronExpression(Cron.AtMinute2)));
+
+        await StartNCronJob(startMonitoringEvents: true);
+
+        var registry = ServiceProvider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var jobRegistry = ServiceProvider.GetRequiredService<JobRegistry>();
+        jobRegistry.FindAllJobDefinition(typeof(DummyJob)).Count.ShouldBe(2);
+
+        registry.DisableJob<DummyJob>();
+        registry.RemoveJob<DummyJob>();
+
+        var completedOrchestrationEvents = await WaitForNthOrchestrationState(
+            ExecutionState.OrchestrationCompleted,
+            2,
+            stopMonitoringEvents: true);
+
+        jobRegistry.FindAllJobDefinition(typeof(DummyJob)).ShouldBeEmpty();
+
+        var firstOrchestrationEvents = Events.FilterByOrchestrationId(completedOrchestrationEvents[0].CorrelationId);
+        firstOrchestrationEvents.ShouldBeScheduledThenCancelled();
+
+        var secondOrchestrationEvents = Events.FilterByOrchestrationId(completedOrchestrationEvents[1].CorrelationId);
+        secondOrchestrationEvents.ShouldBeScheduledThenCancelled();
+    }
+
+    [Fact]
     public async Task CanUpdateScheduleOfAJob()
     {
         ServiceCollection.AddNCronJob(s => s.AddJob<DummyJob>(p => p.WithCronExpression("0 0 * * *").WithName("JobName")));